Changes needed for windows drivers

+++ picohttpparser.h
@@ -27,7 +27,6 @@
 #ifndef picohttpparser_h
 #define picohttpparser_h
 
-#include <stdint.h>
 #include <sys/types.h>
 
 #ifdef _MSC_VER
@@ -65,8 +64,8 @@ struct phr_chunked_decoder {
     char consume_trailer;       /* if trailing headers should be consumed */
     char _hex_count;
     char _state;
-    uint64_t _total_read;
-    uint64_t _total_overhead;
+    unsigned long long _total_read;
+    unsigned long long _total_overhead;
 };
 
 /* the function rewrites the buffer given as (buf, bufsz) removing the chunked-
